+++ apps/docs/src/app/core/component-docs/loading-spinner-docs/examples/loading-spinner-example.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-loading-spinner-example',
-    templateUrl: './loading-spinner-example.component.html'
-})
-export class LoadingSpinnerExampleComponent {}