+++ apps/docs/src/app/platform/component-docs/platform-forms/switch/switch-examples/switch-config-example/switch-config-example.component.ts
@@ -1,0 +1,23 @@
+import { Component } from '@angular/core';
+import { FormControl, FormGroup } from '@angular/forms';
+import { SwitchConfig } from '@fundamental-ngx/platform';
+
+export const switchConfigFactory = SwitchConfig.createProviderFactory({
+    contentDensity: 'compact'
+});
+
+export const customSwitchConfigProvider = {
+    provide: SwitchConfig,
+    useFactory: switchConfigFactory
+};
+
+@Component({
+    selector: 'fdp-switch-config-example',
+    templateUrl: './switch-config-example.component.html',
+    providers: [customSwitchConfigProvider]
+})
+export class SwitchConfigExampleComponent {
+    customForm = new FormGroup({
+        switch: new FormControl(false)
+    });
+}