+++ libs/core/src/lib/loading-spinner/public_api.ts
@@ -1,2 +1,0 @@
-export * from './loading-spinner.module';
-export * from './loading-spinner.component';