+++ libs/platform/src/lib/components/form/switch/switch.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { SwitchModule } from '@fundamental-ngx/core';
+import { SwitchComponent } from './switch';
+
+@NgModule({
+    declarations: [SwitchComponent],
+    imports: [CommonModule, FormsModule, SwitchModule],
+    exports: [SwitchComponent]
+})
+export class PlatformSwitchModule {}